requirements.txt
openai
python-dotenv
azure-identity
azure-keyvault-secrets
moviepy
streamlit
requests
pillow
python-dotenv
